Move grapiserver.Server into internal
