+++ pkg/grapiserver/gateway.go
@@ -8,13 +8,14 @@ import (
 	"time"
 
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
+	"github.com/izumin5210/grapi/pkg/grapiserver/internal"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/grpclog"
 )
 
 // NewGatewayServer creates GrpcServer instance.
-func NewGatewayServer(c *Config) Server {
+func NewGatewayServer(c *Config) internal.Server {
 	return &GatewayServer{
 		Config: c,
 	}
